+++ services/sync-client/src/sync/index.js
@@ -58,7 +58,7 @@ function run () {
   cleanUp()
   profiler.log('Clean up: {dt}\n')
 
-  profiler.log(`Total: {t}\n\n`)
+  profiler.log('Total: {t}\n\n')
 
   console.info(getTotals())